refactor: keep only one feature flag
