+++ src/app/Scenes/Fair/Fair.tests.tsx
@@ -244,7 +244,7 @@ describe("Fair", () => {
 
       it("should not be rendered", () => {
         const { queryByLabelText } = renderWithWrappers(<TestRenderer />)
-        expect(queryByLabelText("Search images")).toBeNull()
+        expect(queryByLabelText("Search by image")).toBeNull()
       })
     })
